feat(sdk): make managed readiness and reconnects reliable (#30)
Clients previously collapsed unavailable managed services into
disabled capability flags. Applications could not distinguish an
unsupported feature from one that was still starting or temporarily
unavailable, nor wait for required services to become ready.

Expose typed service status, readiness reasons, aggregate readiness,
and bounded wait APIs consistently across Rust, Python, and
TypeScript. Preserve unavailable services in capability responses so
applications can report useful state instead of treating them as
missing.

Cached producers could outlive their connection after a server
restart, leaving later publishes stuck on dead transport state. Bound
publish attempts, discard failed producer cells, restore authenticated
connections after retryable failures, and validate retry timing
against the current Apache Iggy SDK APIs.

Allow explicit certificate authorities to enable TLS for local and
private endpoints. Add real restart and three-node failover coverage
across Rust, Python, and TypeScript. The same client handle now
recovers after follower and primary restarts using the published
Apache Iggy SDK fixes, with no skipped failover tests.

Finalize fenced lease ownership around explicit holder and delegated
subject identities so acquisition, renewal, release, and fenced
writes share one unambiguous authorization model.

[wire-break]

+++ bench/src/doctor.rs
@@ -192,6 +192,7 @@ fn uds_probe() -> Result<bool, BenchError> {
     Ok(listener.is_ok())
 }
 
+#[allow(clippy::useless_conversion)] // `statvfs` field widths differ between Linux and macOS.
 fn available_disk_bytes(path: &Path) -> Result<u64, BenchError> {
     let path = std::ffi::CString::new(path.as_os_str().as_bytes()).map_err(|_| {
         BenchError::Invalid(format!("disk probe path `{}` contains NUL", path.display()))
@@ -206,7 +207,7 @@ fn available_disk_bytes(path: &Path) -> Result<u64, BenchError> {
     }
     // SAFETY: a successful `statvfs` call initialized the output structure.
     let stats = unsafe { stats.assume_init() };
-    Ok(stats.f_bavail.saturating_mul(stats.f_frsize))
+    Ok(u64::from(stats.f_bavail).saturating_mul(u64::from(stats.f_frsize)))
 }
 
 fn required_disk_bytes(manifest: &SuiteManifest) -> u64 {
